chunked working and form data merging

+++ client.hpp
@@ -32,9 +32,9 @@ class parssingOfBody
 
         parssingOfBody(/* args */);
         void handle_post(client &obj);
-        void handling_chunked_data(std::string &buffer,std::string &headerOfRequest, std::string &boundary,std::string &bodyofRequest, int & total_bytes_received, unsigned long & ContentLength, int & i, int & bytes_received,int & flag_);
+        void handling_chunked_data(std::string &buffer,std::string &headerOfRequest, std::string &boundary,std::string &bodyofRequest, int & total_bytes_received, int & ContentLength, int & i, int & bytes_received,int & flag_);
         // void handling_chunked_data(client &obj);
-        void handling_form_data(std::string& buffer, std::string &headerOfRequest, std::string &boundary,std::string & bodyofRequest ,int &total_bytes_received,unsigned long &ContentLength,  int & i, int & bytes_received, int & flag_);
+        void handling_form_data(std::string& buffer, std::string &headerOfRequest, std::string &boundary,std::string & bodyofRequest ,int &total_bytes_received,int &ContentLength,  int & i, int & bytes_received, int & flag_);
         void putDataTofile(std::string  data, std::string & bodyofRequest);
         void  create_file_and_put_content(std::string & bodyofRequest,std::string & headerOfRequest);
 
@@ -83,7 +83,7 @@ class client
 	int i;
 	int j;
 	int len;
-	unsigned long ContentLength;
+	int ContentLength;
 
 	int extractheader();
 	void openfile();
